# docs/API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BioMed Zones API",
    "version": "2.0.0",
    "description": "Suitability of French territory (Metropole + DROM) for sustainably cultivating 9 medical species. Every value is traceable to a provenance record (`/sources`). Authentication uses httpOnly cookies (SameSite=Strict); state-changing requests from browsers must come from a whitelisted Origin."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Through the nginx gateway"
    },
    {
      "url": "/",
      "description": "Direct"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bz_access",
        "description": "httpOnly access-token cookie set by POST /auth/login (15 min)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "validation_error"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error",
          "message"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/cells": {
      "get": {
        "tags": [
          "Cells"
        ],
        "summary": "Scores of one species over the H3 grid (compact H3 rows or GeoJSON), paginated and cached",
        "description": "format=h3 returns `items` as `[h3, score, category, confidence]` rows (page size up to 50,000, default 25,000 — the whole resolution-6 grid fits in one page). format=geojson returns polygons (page size up to 5,000, default 2,000).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 60,
              "pattern": "^[a-z-]+$"
            },
            "required": true,
            "name": "species",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "default": 6
            },
            "required": false,
            "name": "resolution",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^-?\\d+(\\.\\d+)?(,-?\\d+(\\.\\d+)?){3}$"
            },
            "required": false,
            "name": "bbox",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "minimum": 0,
              "maximum": 100,
              "default": 0
            },
            "required": false,
            "name": "minScore",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "FXX",
                "GLP",
                "MTQ",
                "GUF",
                "REU",
                "MYT"
              ]
            },
            "required": false,
            "name": "territory",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ],
              "default": "false",
              "description": "Drop cells where protected areas cover ≥ 50 % of the cell, and excluded cells"
            },
            "required": false,
            "name": "excludeProtected",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "h3",
                "geojson"
              ],
              "default": "h3"
            },
            "required": false,
            "name": "format",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50000
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/cells/{h3}": {
      "get": {
        "tags": [
          "Cells"
        ],
        "summary": "Full profile of one cell: environment with provenance, and scores for every species",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "h3",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/compare": {
      "get": {
        "tags": [
          "Cells"
        ],
        "summary": "Two to four cells side by side for one species (parameter fits for a radar chart + table)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "cells",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 60,
              "pattern": "^[a-z-]+$"
            },
            "required": true,
            "name": "species",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}